src/requirements.txt
Flask
python-dotenv
azure-cosmosdb-table
azure-storage-blob
uuid